+++ tester.js
@@ -1,9 +1,9 @@
 
 import { MangaSearch, Manga } from './mangaScraper/mangaScraper.js';
-import { Mangatoto } from './mangaScraper/Fetchers/mangatoto.js';
 
+let mangaSearch = new MangaSearch();
 let manga = new Manga();
-let mangatoto = new Mangatoto();
+
 
 //set config
 manga.updateConfig({
@@ -31,4 +31,37 @@ p.innerHTML = titles;
 let img = document.getElementById("2");
 img.src = pictureLinks[1][0];
 
-console.log(await mangatoto.search("hi"));
+mangaSearch.updateConfig({ language : "eng", maxParallelRequests : 5 });
+
+console.log(await mangaSearch.search("hi"));
+
+function test(mangas) {
+    let titles = [];
+    for (let manga of mangas) {
+        titles.push(manga.get("title"));
+    }
+    console.log(titles);
+    return;
+}
+
+await mangaSearch.search("hi", test);
+
+mangaSearch.updateConfig({runCallbackOnError : true});
+
+function testWithRunCallbackOnError(input) {
+    if (input.status === "fulfilled") {
+        let mangas = input.value;
+        let titles = [];
+        for (let manga of mangas) {
+            titles.push(manga.get("title"));
+        }
+        console.log(titles);
+    } else {
+        console.error(input.value);
+    }
+    return;
+
+}
+
+console.log("testWithRunCallbackOnError");
+mangaSearch.search("hi", testWithRunCallbackOnError);